- add realm database - add method to check if Wifi is enabled

+++ app/src/main/java/com/example/doan/shiftworker/ui/ShiftDetailActivity.java
@@ -44,7 +44,7 @@ public class ShiftDetailActivity extends BaseActivity implements ShiftDetailPres
     }
   }
 
-  @Override public void setId(int id) {
+  @Override public void setId(long id) {
     idTxt.setText(String.valueOf(id));
   }
